requirements.txt
Scrapy==1.1.0
python-dateutil==2.5.3